+++ src/main/java/subway/controller/request/StationRequest.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.controller.request;
 
 public class StationRequest {
     private String name;
