+++ src/main/java/subway/controller/resonse/StationResponse.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.controller.resonse;
 
 public class StationResponse {
     private Long id;
